# doc/swagger.yaml
swagger: '2.0'
info:
  title: Paperwork - API Documentation
  description: Paperwork API documentation for Epitech Project.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  contact: 
    name: Paperwork Team
    url: http://eip.epitech.eu/2024/paperwork
    email: paperwork_2024@labeip.epitech.eu
host: 'localhost:8282'
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
...
paths:
  /user/register:
    post:
      parameters:
        - in: body
          name: User
          description: The user to create.
          schema:
            type: object
            required:
              - email
              - username
              - password
            properties:
              email:
                type: string
                example: test@test.test
              username:
                type: string
                example: username
              password:
                type: string
                example: password
              language:
                type: string
                example: english
      responses:
        '200':
          description: User registered.
          schema:
            type: object
            example:
              {
                "message": "User registered !",
                "email": "test@test.com",
                "jwt": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7ImlkIjoyLCJlbWFpbCI6InRlc3RAdGVzdC5jb20ifSwiaWF0IjoxNjcyNzUyMDY2fQ.oAEpSW-SKnhAvmxMnLMVKuMevsRWqF9Y8ar15JNft6U"
              }
        '400':
          description: Pseudo, email or password missing.
        '409':
          description: Username or email already used.
        '500':
          description: System error.
      summary: Register a new user.
      tags:
        - user
  /user/login:
    post:
      parameters:
        - in: body
          name: User
          description: The user to login.
          schema:
            type: object
            required:
              - email
              - password
            properties:
              email:
                type: string
                example: test@test.test
              password:
                type: string
                example: password
      responses:
        '200':
          description: User logged in.
          schema:
            type: object
            example:
              {
                "message": "User logged in !",
                "email": "test@test.test",
                "jwt": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7ImlkIjoyLCJlbWFpbCI6InRlc3RAdGVzdC5jb20ifSwiaWF0IjoxNjcyNzUyMDY2fQ.oAEpSW-SKnhAvmxMnLMVKuMevsRWqF9Y8ar15JNft6U"
              }
        '404':
          description: User does not exist.
        '400':
          description: Email or password missing / invalid email or password.
        '500':
          description: System error.
      summary: Login a user.
      tags:
        - user
  /user/getbyemail:
    get:
      parameters:
        - in: query
          name: email
          description: User email.
          type: string
      responses:
        '200':
          description: User data
          schema:
            type: object
            example:
              {
                "username": "test",
                "email": "test@test.test"
              }
        '400':
          description: Missing parameter email.
        '404':
          description: User not found.
        '500':
          description: System error.
      summary: Get user information by email.
      tags:
          - user
  /user/getbyusername:
    get:
      parameters:
        - in: query
          name: username
          description: User username.
          type: string
      responses:
        '200':
          description: User data.
          schema:
            type: object
            example:
              {
                "username": "test",
                "email": "test@test.test"
              }
        '400':
          description: Missing parameter username.
        '404':
          description: User not found.
        '500':
          description: System error.
      summary: Get user information by username.
      tags:
          - user
  /user/getbytoken:
    get:
      parameters:
        - in: query
          name: token
          type: string
          description: User token.
      responses:
        '200':
          description: User data.
          schema:
            type: object
            example:
              {
                "id": 238,
                "username": "test",
                "email": "test@test.test",
                "password": "$2a$10$SE.uo7JNlc8k6aLFuEO6LupENW3MKZ7L4E5kPJLCV5aK4B/b7RQU6",
                "language": null,
                "name": null,
                "firstname": null,
                "adress": null,
                "profile_picture": null,
                "age": null,
                "number_phone": null,
                "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7ImlkIjoyMzgsImVtYWlsIjoidGVzdEB0ZXN0LnRlc3QifSwiaWF0IjoxNjgwMDczMTkwfQ.h-gh-2xjS4KAb_TBVXpQEJZ10K6WF3Jyq5olROTyIJs",
                "google_token": null,
                "facebook_token": null
              }
        '400':
          description: Missing parameter token.
        '404':
          description: User not found.
        '500':
          description: System error.
      summary: Get all user information by token.
      tags:
          - user
  /user/delete:
    get:
      parameters:
        - in: query
          name: token
          type: string
          description: User token.
      responses:
        '200':
          description: User deleted
        '400':
          description: Missing parameter token.
        '404':
          description: User not found.
        '500':
          description: System error.
      summary: Delete user.
      tags:
          - user
  /user/modifyDatas:
    post:
      parameters:
        - in: body
          name: User data
          description: The user data to modify.
          schema:
            type: object
            required:
              - token
            properties:
              token:
                type: string
                example: userTokenExemple
              new_email:
                type: string
                example: test@test.test
              password:
                type: string
                example: newPassword
              username:
                type: string
                example: newUsername
              firstname:
                type: string
                example: Emma
              name:
                type: string
                example: lastName
              address:
                type: string
              profile_picture:
                type: string
              number_phone:
                type: integer
                example: 0123456789
      responses:
        '200':
          description: User updated.
        '400':
          description: Missing parameter token.
        '404':
          description: User not found.
        '409':
          description: Email or username already used.
        '500':
          description: System error.
      summary: Modify user data.
      tags:
          - user
  /user/modifySettings:
    get:
      parameters:
        - in: query
          name: token
          type: string
          description: User token.
        - in: query
          name: night_mode
          type: boolean
          description: User preference ( true = night mode, false = light mode).
      responses:
        '200':
          description: User settings updated.
        '400':
          description: Missing parameter token.
        '404':
          description: User not found.
        '500':
          description: System error.
      summary: Update user's settings.
      tags:
          - user
  /user/sendVerificationEmail:
    get:
      parameters:
        - in: query
          name: token
          type: string
          description: User token.
      responses:
        '200':
          description: Email sent.
        '400':
          description: Missing parameter token.
        '404':
          description: User not found.
        '500':
          description: System error.
      summary: Send verification email.
      tags:
          - user
  /user/verifyEmail:
    get:
      parameters:
        - in: query
          name: token
          type: string
          description: User token.
        - in: query
          name: code
          type: string
          description: Verification code.
      responses:
        '200':
          description: Email verified.
        '400':
          description: Missing parameter token or code.
        '404':
          description: User not found.
        '500':
          description: System error.
      summary: Verify email.
      tags:
          - user
  /user/sendResetPasswordEmail:
    get:
      parameters:
        - in: query
          name: email
          type: string
          description: User email.
      responses:
        '200':
          description: Email sent.
          schema:
            type: object
            properties:
              token:
                type: string
                description: User token.
        '400':
          description: Missing parameter email.
        '404':
          description: User not found.
        '500':
          description: System error.
      summary: Send reset password email.
      tags:
          - user
  /user/resetPassword:
    get:
      parameters:
        - in: query
          name: token
          type: string
          description: User token.
        - in: query
          name: password
          type: string
          description: New password.
      responses:
        '200':
          description: Password reset.
        '400':
          description: Missing parameter token, or password.
        '404':
          description: User not found.
        '500':
          description: System error.
      summary: Reset password.
      tags:
          - user
...